requirements.txt
MetaTrader5
python-telegram-bot==20.3
PyYAML

